LayoutTests/imported/w3c/web-platform-tests/css/css-transitions/starting-style-cascade-expected.txt


PASS Overridden @starting-style - order of appearance
FAIL @starting-style with higher specificity assert_equals: Transition of color expected "rgb(0, 192, 0)" but got "rgb(0, 128, 0)"
PASS Starting style does not inherit from parent starting style
FAIL Starting style inheriting from parent's after-change style assert_equals: Transition started from parent's after-change style color expected "rgb(0, 192, 0)" but got "rgb(0, 255, 0)"
FAIL Starting style inheriting from parent's after-change style while parent transitioning assert_equals: Parent transition started expected "rgb(0, 64, 0)" but got "rgb(0, 128, 0)"
